// web-console/src/lib/services/manager/models/NewPipelineResponse.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PipelineId } from './PipelineId'
import type { Version } from './Version'

/**
 * Response to a pipeline creation request.
 */
export type NewPipelineResponse = {
  pipeline_id: PipelineId
  version: Version
}